requirements.txt
chromadb==0.6.3
duckduckgo_search==5.3.1b1
hydra-core==1.3.2
langchain==0.3.22
langgraph==0.3.22
langgraph-api==0.0.48
langgraph-cli==0.1.82
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-chroma==0.2.4
langchain-community==0.3.24
openai==1.70.0
python-dotenv==1.1.0
pypdf==5.5.0